+++ weblate/settings_docker.py
@@ -1113,6 +1113,7 @@ CHECK_LIST = [
     "weblate.checks.chars.EndEllipsisCheck",
     "weblate.checks.chars.EndSemicolonCheck",
     "weblate.checks.chars.MaxLengthCheck",
+    "weblate.checks.chars.MultipleCapitalCheck",
     "weblate.checks.chars.KashidaCheck",
     "weblate.checks.chars.PunctuationSpacingCheck",
     "weblate.checks.chars.KabyleCharactersCheck",